[improvement](chat|common|headless|launcher|webapp) Code upgrades and MyBatis-Plus migration
Migrate XML-based MyBatis mappers to annotation-based MyBatis-Plus
across chat, common, and headless modules. Upgrade code quality with
deprecated API replacements, JSqlParser/Calcite improvements, and
HanLP dictionary optimizations. Update frontend UI, build configs,
demo data, and documentation.

+++ chat/server/src/main/java/com/tencent/supersonic/chat/server/persistence/dataobject/ChatQueryDO.java
@@ -7,15 +7,19 @@ import lombok.Data;
 
 import java.util.Date;
 
+/**
+ * 对话查询数据对象 对应表: s2_chat_query
+ */
 @Data
 @TableName("s2_chat_query")
 public class ChatQueryDO {
-    /** */
+
+    /** 问题ID */
     @TableId(type = IdType.AUTO)
     private Long questionId;
 
-    /** */
-    private Integer agentId;
+    /** 智能体ID */
+    private Long agentId;
 
     /** */
     private Date createTime;
